Add files via upload
Bike Sales - Excel Dashboard

This project is about analyzing the Bike sales data to identify the factors that can influence bike sales.
The tool used for this project – Microsoft Excel
Approach:
Data Cleaning and Data Manipulation
•	Checked the datatype of each column
•	Removed duplicate values from the data
•	Removed rows that had missing values
•	Replaced values in the columns like Gender and Marital Status for convenience
•	Added a column named “Age Bracket”, by using the Nested IF condition

Insights:
1.	Age and Gender wise analysis:
o	It was observed that Middle-aged customers around (31-54) purchased more bikes as compared to older and adolescent customers.
o	It was found that Men were earning high as compared to women, and customers with higher salaries were more likely to purchase bikes.
2.	Geography-wise analysis:
o	It is observed that the average income of the Europe region is less as compared to the Pacific and North America.
o	North America had the highest number of bike purchases as compared to the Pacific and Europe regions.
3.	Commute Distance Analysis:
o	There was a higher probability of customers buying bikes if their commuting distance ranged from 0 to 5 miles.
o	It was observed that there was a decline in the purchase of bikes for customers over 10 miles of commute distance.
4.	Education-wise analysis:
o	It was observed that bachelors who have to travel between 0-5 Miles purchased more bikes as compared to the Graduate Degree, High School, Partial College, and Partial High School.
5.	Children-wise analysis:
o	It was observed that individuals without children are more inclined to buy bikes compared to those who have children.
6.	Car Ownership Analysis:
o	Customers in the middle age range with two cars were less likely to purchase bikes.
o	Customers with four cars only purchased bikes when they had to commute a distance of more than 2 miles. Otherwise, the majority did not bother to buy bikes.
o	It was observed that customers with 2 children and one car were more inclined to buy bikes compared to others.
To increase the sales of the bikes the company should target Middle-aged customers, they should offer seasonal discounts so that women could also buy the bike. As customers who had to commute 0-5 miles were more likely to purchase bikes, the company should target customers and offer discounts.
Companies should also focus on college and high school students to increase their sales. As customers without children were less likely to purchase bikes, they should offer discounts or offers for customers owning more children as this would generate profit for the company.